+++ internal/types/gatewaytypes.go
@@ -7,5 +7,26 @@ import (
 type (
 	Gateway              = gatewayv1.Gateway
 	GatewayAddress       = gatewayv1.GatewayAddress
+	GatewaySpec          = gatewayv1.GatewaySpec
 	GatewayStatusAddress = gatewayv1.GatewayStatusAddress
+	Listener             = gatewayv1.Listener
+	HTTPRoute            = gatewayv1.HTTPRoute
+	HTTPRouteSpec        = gatewayv1.HTTPRouteSpec
+	HTTPRouteList        = gatewayv1.HTTPRouteList
+	ParentReference      = gatewayv1.ParentReference
+	CommonRouteSpec      = gatewayv1.CommonRouteSpec
+	Kind                 = gatewayv1.Kind
+	Group                = gatewayv1.Group
+	AllowedRoutes        = gatewayv1.AllowedRoutes
+	RouteGroupKind       = gatewayv1.RouteGroupKind
+	RouteNamespaces      = gatewayv1.RouteNamespaces
+	ObjectName           = gatewayv1.ObjectName
+)
+
+const (
+	HTTPProtocolType = gatewayv1.HTTPProtocolType
+
+	NamespacesFromAll      = gatewayv1.NamespacesFromAll
+	NamespacesFromSame     = gatewayv1.NamespacesFromSame
+	NamespacesFromSelector = gatewayv1.NamespacesFromSelector
 )